requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
attrs==25.3.0
betterproto==2.0.0b7
black==25.1.0
blockbuster==1.5.24
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
fireworks-ai==0.16.1
forbiddenfruit==0.1.4
frozenlist==1.6.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
hyperframe==6.1.0
idna==3.10
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-core==0.3.59
langchain-fireworks==0.3.0
langchain-openai==0.3.16
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-api==0.2.21
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.11
langgraph-sdk==0.1.66
langsmith==0.3.42
MarkupSafe==3.0.2
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
openai==1.78.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
-e git+https://github.com/dldydtjs2965/bigquery-analytics-agent.git@7fba46366e46d56835dda66a9d405c03d36781ea#egg=react_agent&subdirectory=app
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
